refactor(library/Shortfin): renames "**/Body/Batched.ts" to "**/Body/definition.ts"
