+++ src/library/Shortfin/TextToImage/SDXL/Client/Request/Body/definition.ts
