+++ src/library/Shortfin/TextToImage/SDXL/Client/Request/Body/exports.ts
@@ -1,3 +1,3 @@
 export type {
   Shortfin_TextToImage_SDXL_Client_Request_Body_Batched as Batched,
-} from './Batched';
+} from './definition';
